+++ render/actions/issues.actions.js
@@ -6,6 +6,7 @@ import request, { notify } from './helper';
 export const ISSUES_GET_PAGE = 'ISSUES_GET_PAGE';
 export const ISSUES_GET = 'ISSUES_GET';
 export const ISSUES_COMMENTS_SEND = 'ISSUES_COMMENTS_SEND';
+export const ISSUES_COMMENTS_UPDATE = 'ISSUES_COMMENTS_UPDATE';
 export const ISSUES_RESET_SELECTION = 'ISSUES_RESET_SELECTION';
 export const ISSUES_TIME_ENTRY_GET = 'ISSUES_TIME_ENTRY_GET';
 
@@ -58,8 +59,6 @@ const get = id => (dispatch) => {
 };
 
 const sendComments = (issueId, comments) => (dispatch, getState) => {
-  const { user = {} } = getState();
-
   dispatch(notify.start(ISSUES_COMMENTS_SEND, { subject: 'comments' }));
 
   return request({
@@ -71,13 +70,52 @@ const sendComments = (issueId, comments) => (dispatch, getState) => {
     },
     method: 'PUT'
   }).then(() => {
+    return request({
+      url: `/issues/${issueId}.json`,
+      query: {
+        include: 'journals'
+      },
+      id: `getIssueJournalsDetails:${issueId}`
+    }).then(({ data }) => {
+      dispatch(notify.ok(
+        ISSUES_COMMENTS_SEND,
+        data.issue.journals,
+        {subject: 'comments'}
+      ));
+    })
+      .catch((error) => {
+        console.error(`Error when trying to fetch the issue with id ${issueId} after sending comments:`, error.message);
+        dispatch(notify.nok(ISSUES_COMMENTS_SEND, error, { subject: 'comments' }));
+        return error;
+      });
+  }).catch((error) => {
+    console.error(`Error when trying to assign the issue with id ${issueId}:`, error.message);
+    dispatch(notify.nok(ISSUES_COMMENTS_SEND, error, { subject: 'comments' }));
+    return error;
+  });
+};
+
+const updateComments = (issueId, commentId, comments) => (dispatch, getState) => {
+  const { user = {} } = getState();
+
+  dispatch(notify.start(ISSUES_COMMENTS_UPDATE, { subject: 'comments' }));
+
+  return request({
+    url: `/journals/${commentId}.json`,
+    data: {
+      journal: {
+        notes: comments
+      }
+    },
+    method: 'PUT'
+  }).then(() => {
     dispatch(
       notify.ok(
-        ISSUES_COMMENTS_SEND,
+        ISSUES_COMMENTS_UPDATE,
         {
           created_on: moment().toLocaleString(),
           details: [],
-          id: Date.now(),
+          id: commentId,
           notes: comments,
           private_notes: false,
           user: {
@@ -89,8 +127,8 @@ const sendComments = (issueId, comments) => (dispatch, getState) => {
       ));
   })
     .catch((error) => {
-      console.error(`Error when trying to assign the issue with id ${issueId}:`, error.message);
-      dispatch(notify.nok(ISSUES_COMMENTS_SEND, error, { subject: 'comments' }));
+      console.error(`Error when trying to update the journal ${commentId} of the issue with id ${issueId}:`, error.message);
+      dispatch(notify.nok(ISSUES_COMMENTS_UPDATE, error, { subject: 'comments' }));
       return error; // To be consumed by MarkdownEditor to avoid cleaning the comment
     });
 };
@@ -128,5 +166,6 @@ export default {
   get,
   getTimeEntriesPage,
   sendComments,
+  updateComments,
   resetSelected
 };