+++ render/components/SettingsPage/__tests__/OptionsBlock.spec.jsx
@@ -12,7 +12,7 @@ import {
   SETTINGS_UI_STYLE,
   SETTINGS_IDLE_BEHAVIOR,
   SETTINGS_PROGRESS_SLIDER,
-  SETTINGS_IDLE_TIME_DISCARD, SETTINGS_ISSUE_ALWAYS_EDITABLE, SETTINGS_TIMER_CHECKPOINT,
+  SETTINGS_IDLE_TIME_DISCARD, SETTINGS_ISSUE_ALWAYS_EDITABLE, SETTINGS_TIMER_CHECKPOINT, SETTINGS_COMMENTS_EDITABLE,
 } from '../../../actions/settings.actions';
 import OptionsBlock from '../OptionsBlock';
 import { getInstance, reset, initialize } from '../../../../common/request';
@@ -126,6 +126,40 @@ describe('SummaryPage => OptionsBlock component', () => {
     });
   });
 
+  it('should set use of comments editable when select option is chosen', () => {
+    const state = {
+      settings: {
+        showAdvancedTimerControls: false,
+        showClosedIssues: true,
+        isIssueAlwaysEditable: false,
+        areCommentsEditable: false,
+      },
+      user: {
+        id: 1,
+        redmineEndpoint: 'https://redmine.redmine'
+      }
+    };
+    const store = mockStore(state);
+    const wrapper = mount(
+      <Provider store={store}>
+        <ThemeProvider theme={theme}>
+          <OptionsBlock />
+        </ThemeProvider>
+      </Provider>
+    );
+
+    wrapper.find('input[type="checkbox"]').at(2).simulate('change', { target: { checked: true } });
+    expect(store.getActions().length).toBe(1);
+    expect(store.getActions()[0]).toEqual({
+      type: SETTINGS_COMMENTS_EDITABLE,
+      data: {
+        userId: state.user.id,
+        redmineEndpoint: state.user.redmineEndpoint,
+        areCommentsEditable: true
+      }
+    });
+  });
+
   it('should set use of idle behavior when select option is chosen', () => {
     const state = {
       settings: {
@@ -161,7 +195,7 @@ describe('SummaryPage => OptionsBlock component', () => {
       }
     });
 
-    wrapper.find('input[type="checkbox"]').at(2).simulate('change', { target: { checked: true } });
+    wrapper.find('input[type="checkbox"]').at(3).simulate('change', { target: { checked: true } });
     expect(store.getActions().length).toBe(2);
     expect(store.getActions()[1]).toEqual({
       type: SETTINGS_IDLE_TIME_DISCARD,
@@ -183,7 +217,7 @@ describe('SummaryPage => OptionsBlock component', () => {
       }
     });
 
-    expect(wrapper.find('input[type="checkbox"]').at(2).props().checked).toBe(false);
+    expect(wrapper.find('input[type="checkbox"]').at(3).props().checked).toBe(false);
 
     selects.at(0).instance().selectOption({ label: 'Pause if idle for 5 minutes', value: '5m' });
     expect(store.getActions().length).toBe(4);